Final commit. Unfortunately still unable to get the Add candidate to party endpoint to work. Will work on it as separate project over holidays and hope I can get everything done before exam.

+++ src/main/java/localelection/programming24hr/repositories/CandidateRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 @Repository
 public interface CandidateRepository extends JpaRepository<Candidate, Integer> {
     Candidate findCandidateById(Long id);
-    List<Candidate> findCandidateByPartyId(Long id);
+    List<Candidate> findCandidateByPartyId(int id);
     List<Candidate> deleteCandidateById(int id);
     List<Candidate> findCandidateByFirstName(String firstName);
 }
